requirements.txt
easyocr
opencv-python
mysql-connector-python
numpy
cvzone
datetime
date